2/13/23

+++ PythonApplication3.py
@@ -1,6 +1,6 @@
 #Input and output Operations
 
-
+import os
 
 print("thats pretty gay idk")
 print("man this sucks")
@@ -19,7 +19,11 @@ print(workFileFirstLine)
 workFileSecondLine = workFile.readline()
 print(workFileSecondLine)
 workFile.close() #make sure to have this *at the end*
-writeFile = open('urizons.txt',"w")
+
+if os.path.isfile("urizons.txt"):
+    writeFile = open('urizons.txt',"a")
+else:
+    writeFile = open('urizons.txt',"w")
 toLog = input("What do you want to write in the log? ")
-writeFile.toLog
+writeFile.write("\n" + toLog)
 writeFile.close()
